Require Torchlight
I've been debating whether to include Torchlight as a suggestion or a dependency, and settled on the former. While Torchlight is an extension and not a core dependant, I think most users of Hyde are interested in tech and will probably want to use code examples. Installing Torchlight explicitly allows the integration to be tested better, and create a more predictable environment.