Create "hidden" config option for default theme
Hidden (as in not included in default config file) is that if falls under advanced and specific usage. If the default theme is changed, you would need to make sure all Blade views are compatible, and at that point I expect you have already published the views, and can then see the option there.